requirements.txt
beautifulsoup4==4.6.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
parse-torrent-name==1.1.1
SQLAlchemy==1.1.15
alembic==0.9.6
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT-Extended==3.3.1
Flask-Migrate==2.1.1
Flask-ReCaptcha==0.4.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
mock==2.0.0
gunicorn==19.7.1
PyMySQL==0.7.11
pytest==3.2.3
pytest-cov==2.5.1
requests==2.18.1
coveralls==1.2.0
python-coveralls==2.9.1
